+++ packages/cli/src/Interfaces.ts
@@ -1,4 +1,4 @@
-import type { Application, Request, Response } from 'express';
+import type { Application } from 'express';
 import type {
 	ExecutionError,
 	ICredentialDataDecryptedObject,
@@ -22,7 +22,6 @@ import type {
 	FeatureFlags,
 	INodeProperties,
 	IUserSettings,
-	IHttpRequestMethods,
 	StartNodeData,
 } from 'n8n-workflow';
 
@@ -239,34 +238,6 @@ export interface IExternalHooksFunctions {
 	};
 }
 
-export type WebhookCORSRequest = Request & { method: 'OPTIONS' };
-
-export type WebhookRequest = Request<{ path: string }> & {
-	method: IHttpRequestMethods;
-	params: Record<string, string>;
-};
-
-export type WaitingWebhookRequest = WebhookRequest & {
-	params: WebhookRequest['path'] & { suffix?: string };
-};
-
-export interface WebhookAccessControlOptions {
-	allowedOrigins?: string;
-}
-
-export interface IWebhookManager {
-	/** Gets all request methods associated with a webhook path*/
-	getWebhookMethods?: (path: string) => Promise<IHttpRequestMethods[]>;
-
-	/** Find the CORS options matching a path and method */
-	findAccessControlOptions?: (
-		path: string,
-		httpMethod: IHttpRequestMethods,
-	) => Promise<WebhookAccessControlOptions | undefined>;
-
-	executeWebhook(req: WebhookRequest, res: Response): Promise<IResponseCallbackData>;
-}
-
 export interface IVersionNotificationSettings {
 	enabled: boolean;
 	endpoint: string;
@@ -464,13 +435,6 @@ export interface IPushDataWorkerStatusPayload {
 		internal: boolean;
 	}>;
 	version: string;
-}
-
-export interface IResponseCallbackData {
-	data?: IDataObject | IDataObject[];
-	headers?: object;
-	noWebhookResponse?: boolean;
-	responseCode?: number;
 }
 
 export interface INodesTypeData {
