+++ packages/cli/src/webhooks/test-webhook-registrations.service.ts
@@ -3,7 +3,7 @@ import { CacheService } from '@/services/cache/cache.service';
 import type { IWebhookData } from 'n8n-workflow';
 import type { IWorkflowDb } from '@/Interfaces';
 import { TEST_WEBHOOK_TIMEOUT, TEST_WEBHOOK_TIMEOUT_BUFFER } from '@/constants';
-import { OrchestrationService } from './orchestration.service';
+import { OrchestrationService } from '@/services/orchestration.service';
 
 export type TestWebhookRegistration = {
 	pushRef?: string;
