+++ packages/cli/test/integration/active-workflow-manager.test.ts
@@ -8,8 +8,8 @@ import { ActiveWorkflowManager } from '@/ActiveWorkflowManager';
 import { ExternalHooks } from '@/ExternalHooks';
 import { Push } from '@/push';
 import { SecretsHelper } from '@/SecretsHelpers';
-import { WebhookService } from '@/services/webhook.service';
-import * as WebhookHelpers from '@/WebhookHelpers';
+import { WebhookService } from '@/webhooks/webhook.service';
+import * as WebhookHelpers from '@/webhooks/WebhookHelpers';
 import * as AdditionalData from '@/WorkflowExecuteAdditionalData';
 import type { WebhookEntity } from '@db/entities/WebhookEntity';
 import { NodeTypes } from '@/NodeTypes';
